full rebuild
Added interface for the bag
