+++ .gitignore
@@ -1,3 +1,2 @@
-/.phpunit.result.cache
 /vendor
 /.idea
